fix: circular deps
