+++ packages/core/src/embeddings/types.ts
@@ -1,5 +1,5 @@
 import { BaseNode, MetadataMode } from "../Node.js";
-import { TransformComponent } from "../ingestion/index.js";
+import { TransformComponent } from "../ingestion/types.js";
 import { SimilarityType, similarity } from "./utils.js";
 
 const DEFAULT_EMBED_BATCH_SIZE = 10;
