+++ packages/core/src/ingestion/strategies/index.ts
@@ -1,4 +1,5 @@
-import { BaseDocumentStore, VectorStore } from "../../storage/index.js";
+import { BaseDocumentStore } from "../../storage/docStore/types.js";
+import { VectorStore } from "../../storage/vectorStore/types.js";
 import { TransformComponent } from "../types.js";
 import { DuplicatesStrategy } from "./DuplicatesStrategy.js";
 import { UpsertsStrategy } from "./UpsertsStrategy.js";
