+++ packages/core/src/nodeParsers/types.ts
@@ -1,5 +1,5 @@
 import { BaseNode } from "../Node.js";
-import { TransformComponent } from "../ingestion/index.js";
+import { TransformComponent } from "../ingestion/types.js";
 
 /**
  * A NodeParser generates Nodes from Documents
